+++ misc.h
@@ -1,7 +1,7 @@
 #ifndef misc_h_INCLUDED
 #define misc_h_INCLUDED
 
-double sgn(double x) {
+float sgn(float x) {
   if (x < 0) return -1;
   if (x > 0) return +1;
   return 0;
